// lib/schemas/updateCampaignStatusDtoStatus.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * 5Sport API
 * API documentation for 5Sport authentication and services
 * OpenAPI spec version: 1.0
 */

export type UpdateCampaignStatusDtoStatus = typeof UpdateCampaignStatusDtoStatus[keyof typeof UpdateCampaignStatusDtoStatus];


export const UpdateCampaignStatusDtoStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED',
  CANCELLED: 'CANCELLED',
} as const;