// lib/schemas/campaignOrderControllerFindAllPaymentStatus.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * 5Sport API
 * API documentation for 5Sport authentication and services
 * OpenAPI spec version: 1.0
 */

export type CampaignOrderControllerFindAllPaymentStatus = typeof CampaignOrderControllerFindAllPaymentStatus[keyof typeof CampaignOrderControllerFindAllPaymentStatus];


export const CampaignOrderControllerFindAllPaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED',
} as const;